Move ssl exclusive test to its own set of tests
